move utils package into core directory
