+++ core/adapters/eth_tx_test.go
@@ -13,7 +13,7 @@ import (
 	"github.com/smartcontractkit/chainlink/internal/mocks"
 	strpkg "github.com/smartcontractkit/chainlink/store"
 	"github.com/smartcontractkit/chainlink/store/models"
-	"github.com/smartcontractkit/chainlink/utils"
+	"github.com/smartcontractkit/chainlink/core/utils"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
