+++ core/adapters/json_parse.go
@@ -7,9 +7,9 @@ import (
 	"strings"
 
 	simplejson "github.com/bitly/go-simplejson"
+	"github.com/smartcontractkit/chainlink/core/utils"
 	"github.com/smartcontractkit/chainlink/store"
 	"github.com/smartcontractkit/chainlink/store/models"
-	"github.com/smartcontractkit/chainlink/utils"
 )
 
 // JSONParse holds a path to the desired field in a JSON object,
