+++ core/adapters/multiply.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/smartcontractkit/chainlink/store"
 	"github.com/smartcontractkit/chainlink/store/models"
-	"github.com/smartcontractkit/chainlink/utils"
+	"github.com/smartcontractkit/chainlink/core/utils"
 )
 
 // Multiplier represents the number to multiply by in Multiply adapter.
