+++ core/adapters/multiply_sgx.go
@@ -17,7 +17,7 @@ import (
 
 	"github.com/smartcontractkit/chainlink/store"
 	"github.com/smartcontractkit/chainlink/store/models"
-	"github.com/smartcontractkit/chainlink/utils"
+	"github.com/smartcontractkit/chainlink/core/utils"
 )
 
 // Multiplier represents the number to multiply by in Multiply adapter.
