+++ core/adapters/sleep.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/smartcontractkit/chainlink/store"
 	"github.com/smartcontractkit/chainlink/store/models"
-	"github.com/smartcontractkit/chainlink/utils"
+	"github.com/smartcontractkit/chainlink/core/utils"
 )
 
 // Sleep adapter allows a job to do nothing for some amount of wall time.
