+++ core/cmd/local_client.go
@@ -13,7 +13,7 @@ import (
 	strpkg "github.com/smartcontractkit/chainlink/store"
 	"github.com/smartcontractkit/chainlink/store/models"
 	"github.com/smartcontractkit/chainlink/store/presenters"
-	"github.com/smartcontractkit/chainlink/utils"
+	"github.com/smartcontractkit/chainlink/core/utils"
 	clipkg "github.com/urfave/cli"
 	"go.uber.org/zap/zapcore"
 )
