+++ core/cmd/remote_client.go
@@ -18,7 +18,7 @@ import (
 	homedir "github.com/mitchellh/go-homedir"
 	"github.com/smartcontractkit/chainlink/store/models"
 	"github.com/smartcontractkit/chainlink/store/presenters"
-	"github.com/smartcontractkit/chainlink/utils"
+	"github.com/smartcontractkit/chainlink/core/utils"
 	"github.com/smartcontractkit/chainlink/web"
 	"github.com/tidwall/gjson"
 	clipkg "github.com/urfave/cli"
