+++ store/models/user.go
@@ -5,7 +5,7 @@ import (
 	"regexp"
 	"time"
 
-	"github.com/smartcontractkit/chainlink/utils"
+	"github.com/smartcontractkit/chainlink/core/utils"
 )
 
 // User holds the credentials for API user.
